+++ identify_luck_color.py
